Modernize exports
- BREAKING: Expose `detectEthereumProvider` via named export rather than
  default export
- Use ES module syntax rather than arcane TypeScript syntax for exports
- Split up code into files to match our other libraries
