+++ src/window-with-ethereum.ts
@@ -1,0 +1,5 @@
+import { MetaMaskEthereumProvider } from './metamask-ethereum-provider';
+
+export interface WindowWithEthereum {
+  ethereum?: MetaMaskEthereumProvider;
+}
